64-bit build; move all D3D video resources creation/deletion to main thread

+++ OptionsDialog/OptionsPage.cpp
@@ -244,7 +244,7 @@ void OptionsPage::RadioStrMap::LoadConfigVar()
 	{
 		if (_tcsicmp(val, vals[i]) == 0)
 		{
-			intVar = i;
+			intVar = (int)i;
 			break;
 		}
 	}
